progress all support successPercent (#14412)
close #14332
