+++ components/progress/demo/segment.md
@@ -17,9 +17,19 @@ A standard progress bar.
 import { Tooltip, Progress } from 'antd';
 
 ReactDOM.render(
-  <Tooltip title="3 done / 3 in progress / 4 to do">
-    <Progress percent={60} successPercent={30} />
-  </Tooltip>,
+  <div>
+    <Tooltip title="3 done / 3 in progress / 4 to do">
+      <Progress percent={60} successPercent={30} />
+    </Tooltip>
+
+    <Tooltip title="3 done / 3 in progress / 4 to do">
+      <Progress percent={60} successPercent={30} type="circle" />
+    </Tooltip>
+
+    <Tooltip title="3 done / 3 in progress / 4 to do">
+      <Progress percent={60} successPercent={30} type="dashboard" />
+    </Tooltip>
+  </div>,
   mountNode
 );
 ````
