+++ components/progress/progress.tsx
@@ -1,16 +1,12 @@
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import Icon from '../icon';
-import { Circle } from 'rc-progress';
 import classNames from 'classnames';
+import Icon from '../icon';
 import { ConfigConsumer, ConfigConsumerProps } from '../config-provider';
 import { tuple } from '../_util/type';
-
-const statusColorMap: Record<string, string> = {
-  normal: '#108ee9',
-  exception: '#ff5500',
-  success: '#87d068',
-};
+import Line from './Line';
+import Circle from './Circle';
+import { validProgress } from './utils';
 
 const ProgressTypes = tuple('line', 'circle', 'dashboard');
 export type ProgressType = (typeof ProgressTypes)[number];
@@ -37,22 +33,15 @@ export interface ProgressProps {
   size?: ProgressSize;
 }
 
-const validProgress = (progress: number | undefined) => {
-  if (!progress || progress < 0) {
-    return 0;
-  } else if (progress > 100) {
-    return 100;
-  }
-  return progress;
-};
-
 export default class Progress extends React.Component<ProgressProps, {}> {
   static defaultProps = {
-    type: 'line' as ProgressProps['type'],
+    type: 'line',
     percent: 0,
     showInfo: true,
     trailColor: '#f3f3f3',
-    size: 'default' as ProgressSize,
+    size: 'default',
+    gapDegree: 0,
+    strokeLinecap: 'round',
   };
 
   static propTypes = {
@@ -69,6 +58,27 @@ export default class Progress extends React.Component<ProgressProps, {}> {
     gapDegree: PropTypes.number,
     default: PropTypes.oneOf(['default', 'small']),
   };
+
+  renderProcessInfo(prefixCls: string, progressStatus: (typeof ProgressStatuses)[number]) {
+    const { showInfo, format, type, percent, successPercent } = this.props;
+    if (!showInfo) return null;
+
+    let text;
+    const textFormatter = format || (percentNumber => `${percentNumber}%`);
+    const iconType = type === 'circle' || type === 'dashboard' ? '' : '-circle';
+    if (format || (progressStatus !== 'exception' && progressStatus !== 'success')) {
+      text = textFormatter(validProgress(percent), validProgress(successPercent));
+    } else if (progressStatus === 'exception') {
+      text = <Icon type={`close${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
+    } else if (progressStatus === 'success') {
+      text = <Icon type={`check${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
+    }
+    return (
+      <span className={`${prefixCls}-text`} title={typeof text === 'string' ? text : undefined}>
+        {text}
+      </span>
+    );
+  }
 
   renderProgress = ({ getPrefixCls }: ConfigConsumerProps) => {
     const props = this.props;
@@ -97,79 +107,22 @@ export default class Progress extends React.Component<ProgressProps, {}> {
       !('status' in props)
         ? 'success'
         : status || 'normal';
-    let progressInfo;
     let progress;
-    const textFormatter = format || (percentNumber => `${percentNumber}%`);
 
-    if (showInfo) {
-      let text;
-      const iconType = type === 'circle' || type === 'dashboard' ? '' : '-circle';
-      if (format || (progressStatus !== 'exception' && progressStatus !== 'success')) {
-        text = textFormatter(validProgress(percent), validProgress(successPercent));
-      } else if (progressStatus === 'exception') {
-        text = <Icon type={`close${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
-      } else if (progressStatus === 'success') {
-        text = <Icon type={`check${iconType}`} theme={type === 'line' ? 'filled' : 'outlined'} />;
-      }
-      progressInfo = (
-        <span className={`${prefixCls}-text`} title={typeof text === 'string' ? text : undefined}>
-          {text}
-        </span>
-      );
-    }
+    const progressInfo = this.renderProcessInfo(prefixCls, progressStatus);
 
+    // Render progress shape
     if (type === 'line') {
-      const percentStyle = {
-        width: `${validProgress(percent)}%`,
-        height: strokeWidth || (size === 'small' ? 6 : 8),
-        background: strokeColor,
-        borderRadius: strokeLinecap === 'square' ? 0 : '100px',
-      };
-      const successPercentStyle = {
-        width: `${validProgress(successPercent)}%`,
-        height: strokeWidth || (size === 'small' ? 6 : 8),
-        borderRadius: strokeLinecap === 'square' ? 0 : '100px',
-      };
-      const successSegment =
-        successPercent !== undefined ? (
-          <div className={`${prefixCls}-success-bg`} style={successPercentStyle} />
-        ) : null;
       progress = (
-        <div>
-          <div className={`${prefixCls}-outer`}>
-            <div className={`${prefixCls}-inner`}>
-              <div className={`${prefixCls}-bg`} style={percentStyle} />
-              {successSegment}
-            </div>
-          </div>
+        <Line {...this.props} prefixCls={prefixCls}>
           {progressInfo}
-        </div>
+        </Line>
       );
     } else if (type === 'circle' || type === 'dashboard') {
-      const circleSize = width || 120;
-      const circleStyle = {
-        width: circleSize,
-        height: circleSize,
-        fontSize: circleSize * 0.15 + 6,
-      };
-      const circleWidth = strokeWidth || 6;
-      const gapPos = gapPosition || (type === 'dashboard' && 'bottom') || 'top';
-      const gapDeg = gapDegree || (type === 'dashboard' && 75);
       progress = (
-        <div className={`${prefixCls}-inner`} style={circleStyle}>
-          <Circle
-            percent={validProgress(percent)}
-            strokeWidth={circleWidth}
-            trailWidth={circleWidth}
-            strokeColor={strokeColor || statusColorMap[progressStatus]}
-            strokeLinecap={strokeLinecap}
-            trailColor={trailColor}
-            prefixCls={prefixCls}
-            gapDegree={gapDeg}
-            gapPosition={gapPos}
-          />
+        <Circle {...this.props} prefixCls={prefixCls} progressStatus={progressStatus}>
           {progressInfo}
-        </div>
+        </Circle>
       );
     }
