+++ components/progress/utils.ts
@@ -1,0 +1,8 @@
+export function validProgress(progress: number | undefined) {
+  if (!progress || progress < 0) {
+    return 0;
+  } else if (progress > 100) {
+    return 100;
+  }
+  return progress;
+};
