+++ package.json
@@ -68,7 +68,7 @@
     "rc-menu": "~7.4.12",
     "rc-notification": "~3.3.0",
     "rc-pagination": "~1.17.7",
-    "rc-progress": "~2.2.6",
+    "rc-progress": "~2.3.0",
     "rc-rate": "~2.5.0",
     "rc-select": "^8.6.7",
     "rc-slider": "~8.6.3",
